requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
python-dateutil==2.2
gunicorn==18.0
psycopg2
pyopenssl
